Дописаны методы репозиториев, добавлены новые классы запросов

+++ src/main/groovy/factory/first/may/backend/models/Person.groovy
@@ -33,7 +33,7 @@ class Person {
     int age
 
     @JsonIgnore
-    Long idWorkshop
+    int idWorkshop
 
     @NotNull
     @ManyToOne(
